clock: move turbo speed management to libtrx
